Fix OESS port capacity

+++ fimutil/al2s/oess.py
@@ -180,7 +180,7 @@ class OessClient:
                         endpoint['description'] = interface['description']
                         endpoint['device_name'] = interface['node']
                         endpoint['interface_name'] = interface['name']
-                        endpoint['capacity'] = str(float(interface['bandwidth']) / 1000.0)
+                        endpoint['capacity'] =str(int(float(interface['bandwidth'])/1000.0))
                         endpoint['vlan_range'] = vlan_range
                         endpoint['cloud_interconnect_id'] = interface['cloud_interconnect_id']
                         endpoint['cloud_interconnect_type'] = interface['cloud_interconnect_type'] 
@@ -193,7 +193,7 @@ class OessClient:
                             endpoint['description'] = interface['description']
                             endpoint['device_name'] = interface['node']
                             endpoint['interface_name'] = interface['name']
-                            endpoint['capacity'] = str(float(interface['bandwidth'])/1000.0)
+                            endpoint['capacity'] = str(int(float(interface['bandwidth'])/1000.0))
                             endpoint['cloud_interconnect_id'] = interface['cloud_interconnect_id']
                             endpoint['cloud_interconnect_type'] = interface['cloud_interconnect_type']
                             endpoint['cloud_region'] = entity['name']
